requirements.txt
python-telegram-bot==20.8
flask==3.1.1
reportlab==4.4.1
qrcode==8.2
requests==2.32.3
gunicorn==23.0.0
pillow==11.2.1
